adding getters for k8s-interface
